requirements.txt
# Core
numpy>=1.24.0
pandas>=2.0.0

# RAG minimal
faiss-cpu>=1.7.4
sentence-transformers>=2.6.0
chromadb>=0.4.0

# LLM & API
groq>=0.5.0
openai>=1.0.0
python-dotenv>=1.0.0

# Utilities
PyPDF2>=3.0.0
tqdm>=4.66.0